edit bug saat upload bukti pembayaran

+++ controllers/ProofOfPayment.js
@@ -35,9 +35,10 @@ export const uploadProofOfPayment = async (req, res) => {
       return res.status(422).json({ msg: "Invalid Images" });
     if (fileSize > 5000000)
       return res.status(422).json({ msg: "Image must be less than 5MB" });
-
     const filepath = `./public/images/${news.image}`;
-    fs.unlinkSync(filepath);
+    if (news.image) {
+      fs.unlinkSync(filepath);
+    }
 
     file.mv(`./public/images/${fileName}`, (err) => {
       if (err) return res.status(500).json({ msg: err.message });
